+++ src/core/private/_getActivities.ts
@@ -1,6 +1,6 @@
+import { endpointFormatDateTool as _endpointFormatDateTool } from '../../utils/timeutil';
 import { _axiosAccountApi } from '../api/_axiosApi';
 import { Credentials, IActivities } from '../libraries';
-import { endpointFormatDateTool as _endpointFormatDateTool } from '../utils/timeutil';
 
 // # _getActivities
 /** PROVIDE: credentials, startTime string and endTime string THEN GET: a 'Promise<IAccountActivity[]>' */
