--- src/core/utils/timeutil/index.ts
@@ -1,2 +1,0 @@
-export { day, naive } from './naiveDate';
-export { endpointFormatDateTool } from './timeutil';
