+++ src/debugtesting.ts
@@ -2,7 +2,7 @@
 import { _credentialsFactory } from './core/api/_credentialsFactory';
 import { id0 } from './core/libraries';
 import { apiFunctions } from './core/private/_getDataFromApi';
-import { day } from './core/utils/timeutil/';
+import { day } from './utils';
 export const log = console.log;
 const myRefreshToken = 'qd0AJcnxOGOKpXzvqzIbzKwgHy3Rm3XJ0';
